extract duplicate code + move to /nav folder
