+++ frontend/src/assets/stylesheets/navbar.css
@@ -18,6 +18,11 @@
   @apply absolute left-1/2 -translate-x-1/2;
 }
 
+/* ===== NAV LEFT ===== */
+.nav-left { @apply flex items-center; }
+.nav-links { @apply hidden sm:flex items-center gap-6; }
+.logo { @apply h-6 w-auto; }
+
 /* ===== HAMBURGER ===== */
 .hamburger {
   @apply flex sm:hidden flex-col gap-[5px] cursor-pointer p-1;
@@ -43,6 +48,14 @@
 }
 .drawer-link.router-link-active { @apply bg-surface-inv-raised text-ink-on-inv; }
 .drawer-divider { @apply my-1 border-t border-surface-inv-border; }
+
+/* ===== DRAWER PROFILE ===== */
+.drawer-profile { @apply flex items-center gap-3 px-3 py-2; }
+.drawer-avatar-img { @apply h-9 w-9 rounded-full object-cover ring-1 ring-surface-inv-border; }
+.drawer-avatar-fallback {
+  @apply flex h-9 w-9 shrink-0 items-center justify-center rounded-full bg-surface-inv-raised text-xs font-semibold text-ink-on-inv ring-1 ring-surface-inv-border;
+}
+.drawer-username { @apply text-sm font-medium text-ink-on-inv truncate; }
 
 /* ===== DRAWER TRANSITION ===== */
 .drawer-enter-active, .drawer-leave-active { @apply transition-all duration-200 ease-out overflow-hidden; }
