+++ frontend/src/components/nav/AppNavbar.vue
@@ -1,10 +1,7 @@
 <template>
   <nav class="navbar">
-    <!-- Left: hamburger (mobile) + nav links (desktop) -->
     <div class="nav-left">
-      <button class="hamburger" :aria-expanded="menuOpen" aria-label="Toggle menu" @click="menuOpen = !menuOpen">
-        <span /><span /><span />
-      </button>
+      <NavHamburger :open="menuOpen" @click="menuOpen = !menuOpen" />
       <div class="nav-links">
         <RouterLink :to="{ name: RouteNames.HOME, params: { lang: currentLang } }" class="nav-link">
           {{ t("nav.home") }}
@@ -15,34 +12,29 @@
       </div>
     </div>
 
-    <!-- Center: logo -->
     <RouterLink :to="{ name: RouteNames.HOME, params: { lang: currentLang } }" class="navbar-center">
       <img src="@/assets/images/logo.svg" alt="VierNulVier" width="102" height="32" class="logo" />
     </RouterLink>
 
-    <!-- Right: controls (desktop only) -->
     <div class="hidden sm:block">
       <NavControls :is-dark="isDark" @toggle-dark="$emit('toggle-dark')" />
     </div>
   </nav>
 
-  <!-- Mobile drawer -->
-  <Transition name="drawer">
-    <div v-if="menuOpen" class="mobile-drawer">
-      <RouterLink
-        :to="{ name: RouteNames.HOME, params: { lang: currentLang } }"
-        class="drawer-link"
-        @click="menuOpen = false"
-      >{{ t("nav.home") }}</RouterLink>
-      <RouterLink
-        :to="{ name: RouteNames.PRODUCTIONS, params: { lang: currentLang } }"
-        class="drawer-link"
-        @click="menuOpen = false"
-      >{{ t("nav.productions") }}</RouterLink>
-      <div class="drawer-divider" />
-      <NavControls :is-dark="isDark" @toggle-dark="$emit('toggle-dark')" />
-    </div>
-  </Transition>
+  <NavDrawer :open="menuOpen">
+    <RouterLink
+      :to="{ name: RouteNames.HOME, params: { lang: currentLang } }"
+      class="drawer-link"
+      @click="menuOpen = false"
+    >{{ t("nav.home") }}</RouterLink>
+    <RouterLink
+      :to="{ name: RouteNames.PRODUCTIONS, params: { lang: currentLang } }"
+      class="drawer-link"
+      @click="menuOpen = false"
+    >{{ t("nav.productions") }}</RouterLink>
+    <div class="drawer-divider" />
+    <NavControls :is-dark="isDark" @toggle-dark="$emit('toggle-dark')" />
+  </NavDrawer>
 </template>
 
 <script setup lang="ts">
@@ -51,7 +43,9 @@ import { RouterLink } from "vue-router";
 import { useI18n } from "vue-i18n";
 import { i18n, type SupportedLang } from "@/i18n";
 import { RouteNames } from "@/router/routeNames";
-import NavControls from "@/components/NavControls.vue";
+import NavControls from "@/components/nav/NavControls.vue";
+import NavHamburger from "@/components/nav/NavHamburger.vue";
+import NavDrawer from "@/components/nav/NavDrawer.vue";
 import "@/assets/stylesheets/navbar.css";
 
 defineProps<{ isDark: boolean }>();
@@ -61,11 +55,3 @@ const { t } = useI18n();
 const currentLang = computed(() => i18n.global.locale.value as SupportedLang);
 const menuOpen = ref(false);
 </script>
-
-<style scoped>
-@reference "@/style.css";
-
-.logo { @apply h-6 w-auto; }
-.nav-left { @apply flex items-center; }
-.nav-links { @apply hidden sm:flex items-center gap-6; }
-</style>
