+++ frontend/src/components/nav/NavControls.vue
